refactor: move rollups to a diff package
