+++ src/pb/mod.rs
@@ -1,5 +1,12 @@
 // @generated
 pub mod eth {
+    pub mod rollup {
+        // @@protoc_insertion_point(attribute:eth.rollup.v1)
+        pub mod v1 {
+            include!("eth.rollup.v1.rs");
+            // @@protoc_insertion_point(eth.rollup.v1)
+        }
+    }
     pub mod transaction {
         // @@protoc_insertion_point(attribute:eth.transaction.v1)
         pub mod v1 {
